clean structure of MOI_wrapper
